Moved BatchNormalizationQuantizedTanh to common FPGA folder - used by both Quartus and Vivado
